Samarth Saarthi Updated Ui/backend/requirement.txt
fastapi
uvicorn
python-dotenv
google-genai
s